:truck: Rename to better names Register module components and files
