+++ locales/resources/de.json
@@ -115,7 +115,7 @@
   "Sent": "Gesendet",
   "Share": "Teilen",
   "Sign in": "Anmelden",
-  "Sign in now": "Jetzt anmelden",
+  "Continue to Add account": "Jetzt anmelden",
   "Sign out": "Abmelden",
   "Store your passphrase:": "Speichere deine Passphrase:",
   "Tap on the QR Code to share it.": "Tippe auf den QR-Code um ihn zu teilen",
