+++ locales/resources/en.json
@@ -115,7 +115,7 @@
   "Sent": "Sent",
   "Share": "Share",
   "Sign in": "Sign in",
-  "Sign in now": "Sign in now",
+  "Continue to Add account": "Continue to Add account",
   "Sign out": "Sign out",
   "Store your passphrase:": "Store your passphrase:",
   "Tap on the QR Code to share it.": "Tap on the QR Code to share it.",
