+++ src/modules/Auth/Register/RegisterConfirm/styles.js
