+++ src/modules/Auth/Register/RegisterIntro/styles.js
