+++ src/modules/Auth/Register/RegisterSuccess/styles.js
